+++ activestorage/test/controllers/direct_uploads_controller_test.rb
@@ -6,6 +6,8 @@ require "database/setup"
 if SERVICE_CONFIGURATIONS[:s3] && SERVICE_CONFIGURATIONS[:s3][:access_key_id].present?
   class ActiveStorage::S3DirectUploadsControllerTest < ActionDispatch::IntegrationTest
     setup do
+      @config = SERVICE_CONFIGURATIONS[:s3]
+
       @old_service = ActiveStorage::Blob.service
       ActiveStorage::Blob.service = ActiveStorage::Service.configure(:s3, SERVICE_CONFIGURATIONS)
     end
@@ -15,7 +17,7 @@ if SERVICE_CONFIGURATIONS[:s3] && SERVICE_CONFIGURATIONS[:s3][:access_key_id].pr
     end
 
     test "creating new direct upload" do
-      checksum = ActiveStorage.checksum_implementation.base64digest("Hello")
+      checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
       metadata = {
         "foo" => "bar",
         "my_key_1" => "my_value_1",
@@ -28,18 +30,18 @@ if SERVICE_CONFIGURATIONS[:s3] && SERVICE_CONFIGURATIONS[:s3][:access_key_id].pr
       }
 
       post rails_direct_uploads_url, params: { blob: {
-        filename: "hello.txt", byte_size: 6, checksum: checksum, content_type: "text/plain", metadata: metadata } }
+        filename: "hello.txt", byte_size: 6, checksum: checksum.digest, content_type: "text/plain", metadata: metadata } }
 
       response.parsed_body.tap do |details|
         assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
         assert_equal "hello.txt", details["filename"]
         assert_equal 6, details["byte_size"]
-        assert_equal checksum, details["checksum"]
+        assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
         assert_equal metadata, details["metadata"]
         assert_equal "text/plain", details["content_type"]
         assert_match SERVICE_CONFIGURATIONS[:s3][:bucket], details["direct_upload"]["url"]
         assert_match(/s3(-[-a-z0-9]+)?\.(\S+)?amazonaws\.com/, details["direct_upload"]["url"])
-        assert_equal({ "Content-Type" => "text/plain", "Content-MD5" => checksum, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-amz-meta-my_key_3" => "my_value_3" }, details["direct_upload"]["headers"])
+        assert_equal({ "Content-Type" => "text/plain", "Content-MD5" => checksum.digest, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-amz-meta-my_key_3" => "my_value_3" }, details["direct_upload"]["headers"])
       end
     end
   end
@@ -61,7 +63,7 @@ if SERVICE_CONFIGURATIONS[:gcs]
     end
 
     test "creating new direct upload" do
-      checksum = ActiveStorage.checksum_implementation.base64digest("Hello")
+      checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
       metadata = {
         "foo" => "bar",
         "my_key_1" => "my_value_1",
@@ -74,17 +76,17 @@ if SERVICE_CONFIGURATIONS[:gcs]
       }
 
       post rails_direct_uploads_url, params: { blob: {
-        filename: "hello.txt", byte_size: 6, checksum: checksum, content_type: "text/plain", metadata: metadata } }
+        filename: "hello.txt", byte_size: 6, checksum: checksum.digest, content_type: "text/plain", metadata: metadata } }
 
       response.parsed_body.tap do |details|
         assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
         assert_equal "hello.txt", details["filename"]
         assert_equal 6, details["byte_size"]
-        assert_equal checksum, details["checksum"]
+        assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
         assert_equal metadata, details["metadata"]
         assert_equal "text/plain", details["content_type"]
         assert_match %r{storage\.googleapis\.com/#{@config[:bucket]}}, details["direct_upload"]["url"]
-        assert_equal({ "Content-MD5" => checksum, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-goog-meta-my_key_3" => "my_value_3" }, details["direct_upload"]["headers"])
+        assert_equal({ "Content-MD5" => checksum.digest, "Content-Disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-goog-meta-my_key_3" => "my_value_3" }, details["direct_upload"]["headers"])
       end
     end
   end
@@ -106,7 +108,7 @@ if SERVICE_CONFIGURATIONS[:azure]
     end
 
     test "creating new direct upload" do
-      checksum = ActiveStorage.checksum_implementation.base64digest("Hello")
+      checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
       metadata = {
         "foo" => "bar",
         "my_key_1" => "my_value_1",
@@ -116,17 +118,17 @@ if SERVICE_CONFIGURATIONS[:azure]
       }
 
       post rails_direct_uploads_url, params: { blob: {
-        filename: "hello.txt", byte_size: 6, checksum: checksum, content_type: "text/plain", metadata: metadata } }
+        filename: "hello.txt", byte_size: 6, checksum: checksum.digest, content_type: "text/plain", metadata: metadata } }
 
       response.parsed_body.tap do |details|
         assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
         assert_equal "hello.txt", details["filename"]
         assert_equal 6, details["byte_size"]
-        assert_equal checksum, details["checksum"]
+        assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
         assert_equal metadata, details["metadata"]
         assert_equal "text/plain", details["content_type"]
         assert_match %r{#{@config[:storage_account_name]}\.blob\.core\.windows\.net/#{@config[:container]}}, details["direct_upload"]["url"]
-        assert_equal({ "Content-Type" => "text/plain", "Content-MD5" => checksum, "x-ms-blob-content-disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-ms-blob-type" => "BlockBlob" }, details["direct_upload"]["headers"])
+        assert_equal({ "Content-Type" => "text/plain", "Content-MD5" => checksum.digest, "x-ms-blob-content-disposition" => "inline; filename=\"hello.txt\"; filename*=UTF-8''hello.txt", "x-ms-blob-type" => "BlockBlob" }, details["direct_upload"]["headers"])
       end
     end
   end
@@ -136,7 +138,7 @@ end
 
 class ActiveStorage::DiskDirectUploadsControllerTest < ActionDispatch::IntegrationTest
   test "creating new direct upload" do
-    checksum = ActiveStorage.checksum_implementation.base64digest("Hello")
+    checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
     metadata = {
       "foo" => "bar",
       "my_key_1" => "my_value_1",
@@ -146,13 +148,91 @@ class ActiveStorage::DiskDirectUploadsControllerTest < ActionDispatch::Integrati
     }
 
     post rails_direct_uploads_url, params: { blob: {
-      filename: "hello.txt", byte_size: 6, checksum: checksum, content_type: "text/plain", metadata: metadata } }
+      filename: "hello.txt", byte_size: 6, checksum: checksum.digest, content_type: "text/plain", metadata: metadata } }
 
     response.parsed_body.tap do |details|
       assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
       assert_equal "hello.txt", details["filename"]
       assert_equal 6, details["byte_size"]
-      assert_equal checksum, details["checksum"]
+      assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
+      assert_equal metadata, details["metadata"]
+      assert_equal "text/plain", details["content_type"]
+      assert_match(/rails\/active_storage\/disk/, details["direct_upload"]["url"])
+      assert_equal({ "Content-Type" => "text/plain" }, details["direct_upload"]["headers"])
+    end
+  end
+
+  test "creating new direct upload with checksum/algorithm as combined string" do
+    checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
+    metadata = {
+      "foo" => "bar",
+      "my_key_1" => "my_value_1",
+      "my_key_2" => "my_value_2",
+      "platform" => "my_platform",
+      "library_ID" => "12345"
+    }
+
+    post rails_direct_uploads_url, params: { blob: {
+      filename: "hello.txt", byte_size: 6, checksum: ActiveStorage::Checksum.dump(checksum), content_type: "text/plain", metadata: metadata } }
+
+    response.parsed_body.tap do |details|
+      assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
+      assert_equal "hello.txt", details["filename"]
+      assert_equal 6, details["byte_size"]
+      assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
+      assert_equal metadata, details["metadata"]
+      assert_equal "text/plain", details["content_type"]
+      assert_match(/rails\/active_storage\/disk/, details["direct_upload"]["url"])
+      assert_equal({ "Content-Type" => "text/plain" }, details["direct_upload"]["headers"])
+    end
+  end
+
+  test "creating new direct upload with checksum/algorithm as hash format" do
+    checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
+    metadata = {
+      "foo" => "bar",
+      "my_key_1" => "my_value_1",
+      "my_key_2" => "my_value_2",
+      "platform" => "my_platform",
+      "library_ID" => "12345"
+    }
+
+    post rails_direct_uploads_url, params: { blob: {
+      filename: "hello.txt", byte_size: 6, checksum: checksum.as_json, content_type: "text/plain", metadata: metadata } }
+
+    response.parsed_body.tap do |details|
+      assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
+      assert_equal "hello.txt", details["filename"]
+      assert_equal 6, details["byte_size"]
+      assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
+      assert_equal metadata, details["metadata"]
+      assert_equal "text/plain", details["content_type"]
+      assert_match(/rails\/active_storage\/disk/, details["direct_upload"]["url"])
+      assert_equal({ "Content-Type" => "text/plain" }, details["direct_upload"]["headers"])
+    end
+  end
+
+  test "creating new direct upload with checksum/algorithm not matching service" do
+    algorithm = :SHA256
+    assert_not_equal ActiveStorage::Blob.service.checksum_algorithm, algorithm
+
+    checksum = ActiveStorage::Checksum.base64digest("Hello", algorithm)
+    metadata = {
+      "foo" => "bar",
+      "my_key_1" => "my_value_1",
+      "my_key_2" => "my_value_2",
+      "platform" => "my_platform",
+      "library_ID" => "12345"
+    }
+
+    post rails_direct_uploads_url, params: { blob: {
+      filename: "hello.txt", byte_size: 6, checksum: checksum.as_json, content_type: "text/plain", metadata: metadata } }
+
+    response.parsed_body.tap do |details|
+      assert_equal ActiveStorage::Blob.find(details["id"]), ActiveStorage::Blob.find_signed!(details["signed_id"])
+      assert_equal "hello.txt", details["filename"]
+      assert_equal 6, details["byte_size"]
+      assert_equal checksum, ActiveStorage::Checksum.new(details["checksum"]["digest"], details["checksum"]["algorithm"].to_sym)
       assert_equal metadata, details["metadata"]
       assert_equal "text/plain", details["content_type"]
       assert_match(/rails\/active_storage\/disk/, details["direct_upload"]["url"])
@@ -161,7 +241,7 @@ class ActiveStorage::DiskDirectUploadsControllerTest < ActionDispatch::Integrati
   end
 
   test "creating new direct upload does not include root in json" do
-    checksum = ActiveStorage.checksum_implementation.base64digest("Hello")
+    checksum = ActiveStorage::Checksum.base64digest("Hello", ActiveStorage::Blob.service.checksum_algorithm)
     metadata = {
       "foo" => "bar",
       "my_key_1" => "my_value_1",
@@ -172,7 +252,7 @@ class ActiveStorage::DiskDirectUploadsControllerTest < ActionDispatch::Integrati
 
     set_include_root_in_json(true) do
       post rails_direct_uploads_url, params: { blob: {
-        filename: "hello.txt", byte_size: 6, checksum: checksum, content_type: "text/plain", metadata: metadata } }
+        filename: "hello.txt", byte_size: 6, checksum: checksum.digest, content_type: "text/plain", metadata: metadata } }
     end
 
     response.parsed_body.tap do |details|