+++ activestorage/test/service/s3_service_test.rb
@@ -17,14 +17,14 @@ if SERVICE_CONFIGURATIONS[:s3]
     test "direct upload" do
       key      = SecureRandom.base58(24)
       data     = "Something else entirely!"
-      checksum = ActiveStorage.checksum_implementation.base64digest(data)
+      checksum = ActiveStorage::Checksum.base64digest(data, ActiveStorage::Blob.service.checksum_algorithm)
       url      = @service.url_for_direct_upload(key, expires_in: 5.minutes, content_type: "text/plain", content_length: data.size, checksum: checksum)
 
       uri = URI.parse url
       request = Net::HTTP::Put.new uri.request_uri
       request.body = data
       request.add_field "Content-Type", "text/plain"
-      request.add_field "Content-MD5", checksum
+      request.add_field "Content-MD5", checksum.digest
       Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
         http.request request
       end
@@ -34,10 +34,33 @@ if SERVICE_CONFIGURATIONS[:s3]
       @service.delete key
     end
 
+    test "direct upload with SHA256 checksum" do
+      service = build_service(checksum_algorithm: :SHA256)
+
+      key      = SecureRandom.base58(24)
+      data     = "Something else entirely!"
+
+      checksum = ActiveStorage::Checksum.base64digest(data, service.checksum_algorithm)
+      url      = service.url_for_direct_upload(key, expires_in: 5.minutes, content_type: "text/plain", content_length: data.size, checksum: checksum)
+
+      uri = URI.parse url
+      request = Net::HTTP::Put.new uri.request_uri
+      request.body = data
+      request.add_field "Content-Type", "text/plain"
+      request.add_field "x-amz-checksum-sha256", checksum.digest
+      Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
+        http.request request
+      end
+
+      assert_equal data, service.download(key)
+    ensure
+      service.delete key
+    end
+
     test "direct upload with content disposition" do
       key      = SecureRandom.base58(24)
       data     = "Something else entirely!"
-      checksum = ActiveStorage.checksum_implementation.base64digest(data)
+      checksum = ActiveStorage::Checksum.base64digest(data, ActiveStorage::Blob.service.checksum_algorithm)
       url      = @service.url_for_direct_upload(key, expires_in: 5.minutes, content_type: "text/plain", content_length: data.size, checksum: checksum)
 
       uri = URI.parse url
@@ -58,14 +81,14 @@ if SERVICE_CONFIGURATIONS[:s3]
     test "directly uploading file larger than the provided content-length does not work" do
       key      = SecureRandom.base58(24)
       data     = "Some text that is longer than the specified content length"
-      checksum = ActiveStorage.checksum_implementation.base64digest(data)
+      checksum = ActiveStorage::Checksum.base64digest(data, ActiveStorage::Blob.service.checksum_algorithm)
       url      = @service.url_for_direct_upload(key, expires_in: 5.minutes, content_type: "text/plain", content_length: data.size - 1, checksum: checksum)
 
       uri = URI.parse url
       request = Net::HTTP::Put.new uri.request_uri
       request.body = data
       request.add_field "Content-Type", "text/plain"
-      request.add_field "Content-MD5", checksum
+      request.add_field "Content-MD5", checksum.digest
       upload_result = Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
         http.request request
       end
@@ -99,7 +122,7 @@ if SERVICE_CONFIGURATIONS[:s3]
       begin
         key  = SecureRandom.base58(24)
         data = "Something else entirely!"
-        service.upload key, StringIO.new(data), checksum: ActiveStorage.checksum_implementation.base64digest(data)
+        service.upload key, StringIO.new(data), checksum: ActiveStorage::Checksum.base64digest(data, service.checksum_algorithm)
 
         assert_equal "AES256", service.bucket.object(key).server_side_encryption
       ensure
@@ -115,7 +138,7 @@ if SERVICE_CONFIGURATIONS[:s3]
       @service.upload(
         key,
         StringIO.new(data),
-        checksum: ActiveStorage.checksum_implementation.base64digest(data),
+        checksum: ActiveStorage::Checksum.base64digest(data, ActiveStorage::Blob.service.checksum_algorithm),
         filename: "cool_data.txt",
         content_type: content_type
       )
@@ -131,7 +154,7 @@ if SERVICE_CONFIGURATIONS[:s3]
       @service.upload(
         key,
         StringIO.new(data),
-        checksum: Digest::MD5.base64digest(data),
+        checksum: ActiveStorage::Checksum.base64digest(data, ActiveStorage::Blob.service.checksum_algorithm),
         content_type: "text/plain",
         custom_metadata: { "foo" => "baz" },
         filename: "custom_metadata.txt"
@@ -145,6 +168,24 @@ if SERVICE_CONFIGURATIONS[:s3]
       @service.delete key
     end
 
+    test "upload with unsupported checksum" do
+      assert_raises(ActiveStorage::UnsupportedChecksumError) { build_service(checksum_algorithm: :UnknownHashingFunction) }
+    end
+
+    test "upload with SHA256 checksum" do
+      service = build_service(checksum_algorithm: :SHA256)
+
+      begin
+        key  = SecureRandom.base58(24)
+        data = "Something else entirely!"
+        service.upload key, StringIO.new(data), checksum: ActiveStorage::Checksum.base64digest(data, service.checksum_algorithm)
+
+        assert_equal data, service.download(key)
+      ensure
+        service.delete key
+      end
+    end
+
     test "upload with content disposition" do
       key  = SecureRandom.base58(24)
       data = "Something else entirely!"
@@ -152,7 +193,7 @@ if SERVICE_CONFIGURATIONS[:s3]
       @service.upload(
         key,
         StringIO.new(data),
-        checksum: ActiveStorage.checksum_implementation.base64digest(data),
+        checksum: ActiveStorage::Checksum.base64digest(data, ActiveStorage::Blob.service.checksum_algorithm),
         filename: ActiveStorage::Filename.new("cool_data.txt"),
         disposition: :attachment
       )
@@ -169,7 +210,7 @@ if SERVICE_CONFIGURATIONS[:s3]
         key  = SecureRandom.base58(24)
         data = SecureRandom.bytes(8.megabytes)
 
-        service.upload key, StringIO.new(data), checksum: ActiveStorage.checksum_implementation.base64digest(data)
+        service.upload key, StringIO.new(data), checksum: ActiveStorage::Checksum.base64digest(data, service.checksum_algorithm)
         assert data == service.download(key)
       ensure
         service.delete key
@@ -183,7 +224,7 @@ if SERVICE_CONFIGURATIONS[:s3]
         key  = SecureRandom.base58(24)
         data = SecureRandom.bytes(3.megabytes)
 
-        service.upload key, StringIO.new(data), checksum: ActiveStorage.checksum_implementation.base64digest(data)
+        service.upload key, StringIO.new(data), checksum: ActiveStorage::Checksum.base64digest(data, service.checksum_algorithm)
         assert data == service.download(key)
       ensure
         service.delete key