+++ activestorage/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Add support for additional integrity checksum algorithms for S3, GCS, Azure services
+
+    *Matt Pasquini*
+
 *   Delegate `ActiveStorage::Filename#to_str` to `#to_s`
 
     Supports checking String equality:
@@ -11,9 +15,7 @@
 
     *Sean Doyle*
 
-*   Add support for alternative MD5 implementation through `config.active_storage.checksum_implementation`.
-
-    Also automatically degrade to using the slower `Digest::MD5` implementation if `OpenSSL::Digest::MD5`
+    Automatically degrade to using the slower `Digest::MD5` implementation if `OpenSSL::Digest::MD5`
     is found to be disabled because of OpenSSL FIPS mode.
 
     *Matt Pasquini*, *Jean Boussier*